WIP Moved tech news to blogs controller
