+++ config/routes.rb
@@ -11,7 +11,7 @@ Rails.application.routes.draw do
 
   get 'about-me', to: 'pages#about'
   get 'contact', to: 'pages#contact'
-  get 'tech-news', to: 'pages#tech_news'
+  get 'tech-news', to: 'blogs#tech_news'
 
   resources :blogs do 
     member do
